CGFloat precision for MR_colorFromString

+++ MagicalRecord/Categories/DataImport/MagicalImportFunctions.m
@@ -94,9 +94,9 @@ UIColor * MR_colorFromString(NSString *serializedColor)
         return nil;
     }
     
-    UIColor *color = [UIColor colorWithRed:(componentValues[0] / 255.0f)
-                                     green:(componentValues[1] / 255.0f)
-                                      blue:(componentValues[2] / 255.0f)
+    UIColor *color = [UIColor colorWithRed:(componentValues[0] / (CGFloat)255.)
+                                     green:(componentValues[1] / (CGFloat)255.)
+                                      blue:(componentValues[2] / (CGFloat)255.)
                                      alpha:componentValues[3]];
     
     free(componentValues);
@@ -112,10 +112,10 @@ NSColor * MR_colorFromString(NSString *serializedColor)
     {
         return nil;
     }
-  
-    NSColor *color = [NSColor colorWithDeviceRed:(componentValues[0] / 255.0f)
-                                           green:(componentValues[1] / 255.0f)
-                                            blue:(componentValues[2] / 255.0f)
+    
+    NSColor *color = [NSColor colorWithDeviceRed:(componentValues[0] / (CGFloat)255.)
+                                           green:(componentValues[1] / (CGFloat)255.)
+                                            blue:(componentValues[2] / (CGFloat)255.)
                                            alpha:componentValues[3]];
     free(componentValues);
     return color;
